backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Callisto 語音助理後端服務 - 智能語音對話系統"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "faster-whisper>=1.2.1",
    "groq>=1.0.0",
    "matplotlib>=3.10.8",
    "ollama>=0.4.7",
    "onnxruntime>=1.23.2",
    "opencc>=1.2.0",
    "openwakeword>=0.4.0",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-multipart>=0.0.21",
    "python-osc>=1.9.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
